Upgrade Terraform version and add Go module
Terraform was out of date but the way that the Go dependencies were
installed meant that an unreleased version of the core code was
pulled in and the current API for reading the plan was broken.

Note that we'll need to change this code when we upgrade to 0.12 of
Terraform.